Tidy up process handling
