+++ packages/ciphernode/Cargo.toml
@@ -20,6 +20,7 @@ members = [
 
 [workspace.dependencies]
 actix = "0.13.5"
+actix-web = "4.10.2"
 aes-gcm = "0.10.3"
 aggregator = { path = "./aggregator" }
 alloy = { version = "0.5.2", features = ["full", "node-bindings"] }
