+++ packages/ciphernode/enclave_core/Cargo.toml
@@ -35,6 +35,7 @@ test-helpers = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
 zeroize = { workspace = true }
+actix-web = {workspace = true }
 
 [build-dependencies]
 serde_json = { workspace = true }
