+++ packages/ciphernode/enclave_core/src/helpers/mod.rs
@@ -1,4 +1,4 @@
 pub mod datastore;
 pub mod shutdown;
-
+pub mod swarm;
 pub use shutdown::*;
