+++ packages/ciphernode/enclave_core/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod aggregator_start;
-mod helpers;
+pub mod helpers;
 pub mod init;
 pub mod net_generate;
 pub mod net_purge;
@@ -7,7 +7,6 @@ pub mod net_set;
 pub mod password_create;
 pub mod password_delete;
 pub mod start;
+pub mod swarm_daemon;
 pub mod swarm_up;
 pub mod wallet_set;
-
-pub use helpers::*;
